+++ aqueduct/test/http/default_resource_controller_test.dart
@@ -341,10 +341,9 @@ void main() {
       var op = collectionOperations["post"];
       expect(op.id, "createTestModel");
 
-      expect(op.responses.length, 4);
+      expect(op.responses.length, 3);
 
       expect(op.responses["409"], isNotNull);
-      expect(op.responses["422"], isNotNull);
       expect(op.responses["400"], isNotNull);
       expect(op.responses["200"].content["application/json"].schema.referenceURI.path, "/components/schemas/TestModel");
       expect(op.requestBody.content["application/json"].schema.referenceURI.path, "/components/schemas/TestModel");
@@ -354,11 +353,10 @@ void main() {
       var op = idOperations["put"];
       expect(op.id, "updateTestModel");
 
-      expect(op.responses.length, 5);
+      expect(op.responses.length, 4);
 
       expect(op.responses["404"], isNotNull);
       expect(op.responses["409"], isNotNull);
-      expect(op.responses["422"], isNotNull);
       expect(op.responses["400"], isNotNull);
       expect(op.responses["200"].content["application/json"].schema.referenceURI.path, "/components/schemas/TestModel");
       expect(op.requestBody.content["application/json"].schema.referenceURI.path, "/components/schemas/TestModel");