+++ aqueduct_test/lib/src/response.dart
@@ -17,7 +17,7 @@ class TestResponse {
   /// The HTTP body of the response.
   ///
   /// The body is guaranteed to be decoded prior to accessing it. You do
-  /// not need to invoke [TestResponseBody.decodedData] or any of its asynchronous
+  /// not need to invoke [TestResponseBody.decode] or any of its asynchronous
   /// decoding methods.
   final TestResponseBody body;
 
@@ -84,7 +84,7 @@ class TestResponseBody extends BodyDecoder {
 
   @override
   String toString() {
-    return asDynamic().toString();
+    return as().toString();
   }